Official Update: code-style: simplify some checks
main change here is making the  logic saner. the rest of the
changes are just small stuff which accumulated on my local branch.

pop() must not be called with NULL. and  achieves this, but in a
very (unnecessarily) complicated way:

if c == NULL then nexttiled() will return NULL as well, so we enter this
branch:

	if (c == nexttiled(selmon->clients))

in here the cd builds/dwm check fails and the function returns before calling pop()

			return;

however, none of this was needed. we can simply return early if c was NULL.
Also  is set to  so we can use  in the first check
instead which makes things shorter.